commit payment

+++ database/migrations/2022_10_31_161130_create_orders_table.php
@@ -21,7 +21,7 @@ class CreateOrdersTable extends Migration
             $table->string('phone');
             $table->string('email');
             $table->tinyInteger('status');
-            $table->tinyInteger('total_money');
+            $table->integer('total_money');
             $table->timestamps();
             $table->softDeletes();
         });
